to change the parameter order of BiFunction and BinaryOperator taken by heap() variations, replaced them with accum() variations and made Deprecated.
TList is all about list and when handling the prior one and the latter one, 
the parameter order should follow. so does diff()'s parameter, heap()'s parameter
also should. but this time, i reserved heap() untouched for compatibility
consideration, but Deprecated it and newly added accum() as replacement for it.